Include ApplicationName parameter in PrivacyPolicyUrl
As of April 7, 2014 there have been multiple reports of apps failing
submission to the Windows 8 Store on account of the included privacy
policy not being specific to the application.  To overcome this, the
included privacy policy can now provide your ApplicationName into the
example privacy policy provided the 'App' parameter is set.

Example:
http://pjdecarlo.com/2013/11/app-privacy-policy.html?App=INSERT YOUR APP
NAME HERE"

This implies that you will also need to provide the example above, with
your ApplicationName manually added, during the Description step of you
Windows 8 app submission.

Below is a copy of the failure from the Windows 8 Store that you may
encounter your privacy policy is not app-specific:

****************************************************

Content compliance: failed
This test evaluates your software for content compliance with Windows
Store Certification Requirements. Learn more
Your app doesn't meet requirement 4.1. Learn more
Notes from Testers:
The privacy policy provided for this App in the app description page
does not appear to be related to privacy. Note: The privacy policy needs
to be specific to your company and/or App. In addition to your app's
privacy policy, you can also include links to privacy policy of other
service providers if the user's information is collected or shared with
them via this app. The privacy statement provided in the Windows
Settings Charm for this App does not appear to relate to privacy. Note:
The privacy policy needs to be specific to your company and/or App. In
addition to your app's privacy policy, you can also include links to
privacy policy of other service providers if the user's information is
collected or shared with them via this app.